DOMStorage: Add crash keys and use DWoCs when binding storage controls
As part of a recent refactor that removed partition.mojom, we added
CHECKs to ensure that we bind a StorageControl for a given path just the
once. Unfortunately, the issue tracked in crbug.com/396030877 where we
receive multiple Bind requests continues to occur after the refactor.

This change adds crash keys for both BindLocalStorageControl and
BindSessionStorageControl. Additionally, it replaces the CHECKs with
DumpWithoutCrashing.

Bug: 425016325
Change-Id: Ie6bbc84afd9c17f3d961b6e85d696516e284f970
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6646939
Reviewed-by: Evan Stade <evanstade@microsoft.com>
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1485955}