Add gpu workaround for NV12 upload issue
A set of NVidia drivers have an issue where it fails to upload NV12 to a
texture through init data or UpdateSubresource.  MFVEA does this when it
receives an NV12 main memory frame with shared image support turned on,
in order to ensure consistency with the input stream (all textures).
Add a workaround and turn off shared image support for affected drivers.

Bug: 423663681
Change-Id: I100e83951f904c838f3b296e83ec8dd96a5e5aa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6672088
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Matthew Andrews <mattandr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1485992}