[iOS]isActivityViewPresented reflects the view’s state
The value was set to nil before the view was actually dismissed. This
ensure the fake is closer to the real tested object.

This will be particularly useful in a next CL, where we simulate the
view disappearing without the callback being informed.

Bug: 395959814
Change-Id: Ie58ba5add418759f80fd49f0d7dc617a3cc14a98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6722632
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486087}