[RSG] Make AHM::OnGetSingleFieldSuggestions always consume callback
Given the larger effort of refactoring the suggestion generation logic
the new architecture will require AHM::OnGetSingleFieldSuggestions to
own the callback (since the suggestions will be returned from the db
with a delay and AHM won't know in advance whether there are any
suitable entries for the given field).

Because autocomplete entries are generated last, AHM will always call
the callback (possibly with empty vector of suggestions), allowing
other FillingProducts (e.g. PlusAddresses) to be suggested as well.

Removal of the existing call to AHM::CancelPendingQuery() does not
affect the logic since AHM::OnGetSingleFieldSuggestions() calls it too
as one of the first steps.

Bug: 409962888
Change-Id: I2e827b60d1a01111db8f7e11a5d06efce4164ccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6729257
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1486254}