[PasswordChange] Handle hats without password store data
At time of launching hats survey for password change, the data might not
be fetched yet from password store.

HatsService requires all defined product-specific fields to be present
when launching a survey, hence passing empty data is not an option.

UXR suggestion is to pass "-1" for counts in this case, so the survey
answers can still be analyzed and counts ignored.

Fixed: b:429595214
Change-Id: I4b10a1e16c839a61405e40e8e0663752f8a11331
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6733238
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/main@{#1486267}