[PartitionAlloc] Enable priority inheritance earlier
Enabling priority inheritance in partition allocator locks requires all
the locks to migrate from using a non-PI futex to a PI futex. This step
can be slow if lock contention is high. In an earlier version of the
experiment, the finch config was responsible for enabling the feature
but the current version uses cached flags which are available earlier.

This commit enables PI before the thread pool is started to reduce the
probability of slow lock migrations.

Bug: 384902323
Change-Id: If4705358e98aec63e856ce68bd1f216fde581698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6715105
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Anand Ravi <anandrv@google.com>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486333}