Actor: Ensure observed tabs is never empty
This change fixes the DCHECK that would occur when calling
GetTabForObservation after an action. The DCHECK trips because the
observable set of tabs is empty. This can happen if the first action for
a task is not tab-scoped, for example, WaitAction. In that case the
action won't cause the tab to be added to the observable set and thus
the DCHECK is hit.

This CL fixes the issue by checking for the case where the current task
has an empty tab-set and the incoming action won't add one. In that
case, the currently focused tab is added to the set which restores
behavior prior to recent refactorings in https://crbug.com/411462297.

Longer term we may want to decide how to handle cases like this
explicitly.

Bug: 431239173
Change-Id: I7f40c8755d491453b4cefa9f2e662995b5b1407e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731515
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486387}