[iOSPasswordRecovery] Add icon to KA backup password suggestions
This CL adds the `kHistorySymbol` icon to the backup password
suggestions shown in the keyboard accessory. This icon is needed so that
backup password suggestions can be differentiated from their main
passwords.

Given that this exact icon is used for backup suggestions in the
password bottom sheet, keyboard accessory, and soon will also be used in
the manual fallback menu, it felt appropriate to have a util method for
it.

Screenshot: https://screenshot.googleplex.com/76XoDkTpbGaFTBN.png

Mocks:
https://docs.google.com/presentation/d/10Be3B2N3YDgbEporNlPsGLZ5vZZKEBCYq8TQI8h3740/edit?slide=id.g35c2177e5f5_0_508#slide=id.g35c2177e5f5_0_508

Fixed: 417943514
Change-Id: Id042c75e0667b210bc2d5763d5acd50a40fd82dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6673953
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1486417}