Fork api_layer_interface.cpp for libc++ roll
The new libc++ roll requires us to add a direct include to this file.
Since it takes a while for upstream changes to roll in, fork the file
until it's fixed upstream.

The upstream change is
https://github.com/KhronosGroup/OpenXR-SDK-Source/pull/554

For ease of review, the patchsets in this CL are as follows:
1 - copy the file unchanged
2 - Make the change, update the patch and build files

Note that the presubmit complained about formatting, but I wasn't able
to get `git cl format` to actually do anything.

Bug: 431755915
Change-Id: I0a42ccfbb1c95b8494d6590bf691578b82a7074a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6734768
Commit-Queue: Devon Loehr <dloehr@google.com>
Auto-Submit: Devon Loehr <dloehr@google.com>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486440}