extensions: Clean up ExtensionHostDelegate for desktop Android
Start to structure the code like we do for other client interfaces:
- chrome_extension_host_delegate.cc for shared code
- chrome_extension_host_delegate_android.cc for unported android code
- chrome_extension_host_delegate_non_android.cc for other platforms

Also move a couple of methods to the shared file. I'm going to do the
rename of desktop_android_extension_host_dialog.cc in a follow-up so
that git will properly track the rename.

Bug: 415033270
Test: existing unit_tests, browser_test and android_browsertests
Change-Id: Icda3c98664322629d848e1bd8c3079ba67142ae1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6732431
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486474}