[FPF] Add total defer time UMA metric
This adds a histogram that records the total time subresources spend in a deferred state due to throttling on the renderer. The existing metrics "FP.SubresourceLoad.Evaluation.*" are insufficient to measure latency on the renderer because they only measure the time spent matching a URL against the ruleset, not time spent on getting activation state, posting tasks and waiting for them to complete, etc.

Bug: 424582998
Change-Id: I3397077e6f1fe31e74482326d832ab4dcaa76e9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731321
Reviewed-by: Zainab Rizvi <rizvis@google.com>
Commit-Queue: Martin Pan-Verde <thesalsa@google.com>
Cr-Commit-Position: refs/heads/main@{#1486485}