Extension: Allow ChromeVox to load tutorial resources in mv3
This CL is part of the migration of ChromeVox to mv3.
ChromeVox's tutorial is built using Polymer, which imports scripts
via statements like:

import {...} from 'chrome://resources...'

Imports of external and remote scripts are disallowed in mv3, so this
change adds an exception for ChromeVox by modifying its CSP and
updating the CSP validation that occurs when the extension is loaded.

Bug: 388867840
Change-Id: Ia94f304918088497e975f15b05b696fc833cd8dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6719307
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486522}