Refactor: Encapsulate progress bar color logic in Payment Handler UI
Previously, the logic to calculate and apply progress bar colors was
mixed with layout code in PopulateSheetHeaderView(). This made the code
difficult to maintain and led to inefficient UI updates, as the entire
header had to be rebuilt when colors changed.

This patch isolates the color calculation logic by moving it to a new
SetColorBasedOnBackground() method within the PaymentHandlerProgressBar
class. This refactoring improves separation of concerns and unlocks the
path toward more efficient, targeted updates to the header's colors and
text.

Bug: 418001134
Change-Id: I0f4109e54dac57f4e2084960fb0de86b547a89cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6728228
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486538}