Refactor: Encapsulate close button color logic in Payment Handler UI
Previously, logic to apply close button colors was mixed with layout
code in PopulateSheetHeaderView(). This coupling was difficult to
maintain and led to inefficient UI updates, as the entire header had to
be rebuilt for color changes, harming accessibility.

This patch isolates the color logic into a new
SetColorBasedOnBackground() method within the PaymentHandlerCloseButton
class. This refactoring improves separation of concerns and unlocks more
efficient, targeted updates to the header's colors and text.

Bug: 418001134
Change-Id: I82cca94caefc80a5681dc5cf3777b695e28def6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6730034
Reviewed-by: Darwin Yang <darwinyang@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486549}