Clean up TypedStatus using **concepts**
At some point there was a bit of a cleanup to some of the TypedStatus
macros which used concepts & requires. Those changes could have gone
farther, and so I've introduced a whole bunch of concepts stuff to
remove some of the uglier static asserts and weird template magic helper
classes that were stinking up the code. I also removed the unused
MapValue implementation that could change the status type, as it's only
reference was in the test for that function.

R=liberato

Skip-Clang-Tidy-Checks: google-explicit-constructor
Change-Id: I44fcabb37552e3e14e4ba666ef8e588be9c653f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6725199
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486591}