[Extensions] runtime.onMessage() promise return (reject) support.
Before this change runtime.onMessage() listeners who returned a promise
would not keep the message port open for that promise to reject. The
only option was to return `true` from the listener to indicate that the
developer wanted to run asynchronous code before responding (to the
sender's response callback).

After this change a promise may be returned and the message port will
remain open until the promise's rejected value can be provided to the
sender as an error. The port will then be closed and the promise's
error will be sent back as an error to the message sender, consistent
with mozilla/webextensions-polyfill.

I also took the liberty of renaming some of the one time message
functions to have more generic names since they are being used for
message response and port closing behavior.

Bug: 40753031
Change-Id: I729b8c48e44c43913b1c2858aed553d2d7c74787
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6680436
Auto-Submit: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486596}