[PDF Ink Signatures] Refactor PdfCaret::GetScreenRectForChar()
Rename PdfCaret::GetScreenRectForChar() to GetScreenRectForCaret(),
which is a more accurate name given that its implementation modifies the
rect's width to be as wide as a text caret.

This frees up the GetScreenRectForChar() name to be introduced with a
different implementation in https://crrev.com/c/6692495.

Then, remove its parameters, as it is currently always called with the
same member variables passed in.

This is a mechanical change and should have no change in behavior.

Bug: 427133561
Change-Id: I39160013a16a8a902dd08c26afe26d3c58ab21c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6692494
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486651}