[Desktop Android] Move ExtensionTabUtil::GetWindowIdOfTabStripModel()
ExtensionTabUtil::GetWindowIdOfTabStripModel() needs to be removed for
the experimental desktop android platform (where TabStripModel is
unsupported). It has only one callsite, so just move it to be inlined
there. This removes it from ExtensionTabUtil and makes it less likely
that there will be new callers added.

Bug: 430344931
Change-Id: I98a80cf53e454697c9e3069b3218004128d18a29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6735351
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486654}